Update tOctree.h
Added an entry type to pair objects to their current node.
Objects can only exist in 1 node but when the query comes through all nodes that are overlapped by this region check will be returned.
reduced recursivity.
Testing to see the best MAX_OBJECT amount, but this number will be settable in future.