remove all stage specifics and dynamic naming
always running into issues with this as at times features are coming in with a different index and that affects how another stage is interacting with that shader. This produces alot more shaders, previous run with these was around 500 but without it is producing over 800 (this includes tso compiled output)

This may be something that will only work for the shadernodes as users will be designing those.